Initial Stations Search Popup
Built out the initial logic for the weather stations near me lookup. You can now enter your zip code and a popup will show the stations near you (with test data still). Next step is to either add the real endpoint or add functionality to click on a specific station and have it show on a map.